+++ tests/datafiletests.cpp
@@ -18,8 +18,8 @@ enum class Theme
 class AppConfig : public DataFileBase
 {
 public:
-    AppConfig(const std::string& key, const std::string& appName)
-        : DataFileBase{ key, appName } 
+    AppConfig(const std::string& key, const std::string& appName, bool isPortable)
+        : DataFileBase{ key, appName, isPortable }
     { 
 
     }
@@ -85,15 +85,27 @@ class DataFileTest : public testing::Test
 {
 public:
     static std::shared_ptr<DataFileManager> m_manager;
+    static std::shared_ptr<DataFileManager> m_portableManager;
 
     static void SetUpTestSuite()
     {
         std::filesystem::remove(UserDirectories::get(ApplicationUserDirectory::Config, "Nickvision Aura Tests") / ("config.json"));
-        m_manager = std::make_shared<DataFileManager>("Nickvision Aura Tests");
+        std::filesystem::remove("config.json");
+        m_manager = std::make_shared<DataFileManager>("Nickvision Aura Tests", false);
+        m_portableManager = std::make_shared<DataFileManager>("Nickvision Aura Tests", true);
+    }
+
+    static void TearDownTestSuite()
+    {
+        m_manager.reset();
+        m_portableManager.reset();
+        std::filesystem::remove_all(UserDirectories::get(ApplicationUserDirectory::Config, "Nickvision Aura Tests"));
+        std::filesystem::remove("config.json");
     }
 };
 
 std::shared_ptr<DataFileManager> DataFileTest::m_manager{ nullptr };
+std::shared_ptr<DataFileManager> DataFileTest::m_portableManager{ nullptr };
 
 TEST_F(DataFileTest, EnsureDefaultAppConfig)
 {
@@ -106,9 +118,28 @@ TEST_F(DataFileTest, EnsureDefaultAppConfig)
     ASSERT_EQ(config.getAutomaticallyCheckForUpdates(), true);
 }
 
+TEST_F(DataFileTest, EnsureDefaultPortableAppConfig)
+{
+    AppConfig& config{ m_portableManager->get<AppConfig>("config") };
+    WindowGeometry geometry{ config.getWindowGeometry() };
+    ASSERT_EQ(config.getTheme(), Theme::System);
+    ASSERT_EQ(geometry.getWidth(), 800);
+    ASSERT_EQ(geometry.getHeight(), 600);
+    ASSERT_EQ(geometry.isMaximized(), false);
+    ASSERT_EQ(config.getAutomaticallyCheckForUpdates(), true);
+}
+
 TEST_F(DataFileTest, ChangeAppConfig1)
 {
-    AppConfig& config{m_manager->get<AppConfig>("config") };
+    AppConfig& config{ m_manager->get<AppConfig>("config") };
+    ASSERT_NO_THROW(config.setTheme(Theme::Light));
+    ASSERT_NO_THROW(config.setWindowGeometry(WindowGeometry{ 1920, 1080, true }));
+    ASSERT_TRUE(config.save());
+}
+
+TEST_F(DataFileTest, ChangePortableAppConfig1)
+{
+    AppConfig& config{ m_portableManager->get<AppConfig>("config") };
     ASSERT_NO_THROW(config.setTheme(Theme::Light));
     ASSERT_NO_THROW(config.setWindowGeometry(WindowGeometry{ 1920, 1080, true }));
     ASSERT_TRUE(config.save());
@@ -124,9 +155,26 @@ TEST_F(DataFileTest, EnsureChangeInAppConfig)
     ASSERT_EQ(geometry.isMaximized(), true);
 }
 
+TEST_F(DataFileTest, EnsureChangeInPortableAppConfig)
+{
+    AppConfig& config{ m_portableManager->get<AppConfig>("config") };
+    ASSERT_EQ(config.getTheme(), Theme::Light);
+    WindowGeometry geometry{ config.getWindowGeometry() };
+    ASSERT_EQ(geometry.getWidth(), 1920);
+    ASSERT_EQ(geometry.getHeight(), 1080);
+    ASSERT_EQ(geometry.isMaximized(), true);
+}
+
 TEST_F(DataFileTest, ChangeAppConfig2)
 {
     AppConfig& config{ m_manager->get<AppConfig>("config") };
+    ASSERT_NO_THROW(config.setAutomaticallyCheckForUpdates(false));
+    ASSERT_TRUE(config.save());
+}
+
+TEST_F(DataFileTest, ChangePortableAppConfig2)
+{
+    AppConfig& config{ m_portableManager->get<AppConfig>("config") };
     ASSERT_NO_THROW(config.setAutomaticallyCheckForUpdates(false));
     ASSERT_TRUE(config.save());
 }
@@ -137,11 +185,30 @@ TEST_F(DataFileTest, EnsureChangeInAppConfig2)
     ASSERT_EQ(config.getAutomaticallyCheckForUpdates(), false);
 }
 
-TEST_F(DataFileTest, ReloadConfig)
+TEST_F(DataFileTest, EnsureChangeInPortableAppConfig2)
+{
+    AppConfig& config{ m_portableManager->get<AppConfig>("config") };
+    ASSERT_EQ(config.getAutomaticallyCheckForUpdates(), false);
+}
+
+TEST_F(DataFileTest, ReloadAndCheckConfig)
 {
     m_manager.reset();
-    m_manager = std::make_shared<DataFileManager>("Nickvision Aura Tests");
+    m_manager = std::make_shared<DataFileManager>("Nickvision Aura Tests", false);
     AppConfig& config{ m_manager->get<AppConfig>("config") };
+    ASSERT_EQ(config.getTheme(), Theme::Light);
+    WindowGeometry geometry{ config.getWindowGeometry() };
+    ASSERT_EQ(geometry.getWidth(), 1920);
+    ASSERT_EQ(geometry.getHeight(), 1080);
+    ASSERT_EQ(geometry.isMaximized(), true);
+    ASSERT_EQ(config.getAutomaticallyCheckForUpdates(), false);
+}
+
+TEST_F(DataFileTest, ReloadAndCheckPortableConfig)
+{
+    m_portableManager.reset();
+    m_portableManager = std::make_shared<DataFileManager>("Nickvision Aura Tests", true);
+    AppConfig& config{ m_portableManager->get<AppConfig>("config") };
     ASSERT_EQ(config.getTheme(), Theme::Light);
     WindowGeometry geometry{ config.getWindowGeometry() };
     ASSERT_EQ(geometry.getWidth(), 1920);